requirements.txt
annotated-types==0.7.0
anyio==4.8.0
ariadne==0.26.1
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
dnspython==2.7.0
ecdsa==0.19.0
fastapi==0.115.9
graphql-core==3.2.5
h11==0.14.0
idna==3.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
python-jose==3.4.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
vendor~=0.1
PyJWT~=2.10.1
passlib~=1.7.4
dotenv~=0.9.9
python-dotenv~=1.0.1
slowapi~=0.1.9
